support auto syntax for multiple providers
