+++ src/nixpacks/plan/merge.rs
@@ -1,5 +1,6 @@
 use super::{
     phase::{Phase, StartPhase},
+    utils::fill_auto_in_vec,
     BuildPlan,
 };
 
@@ -100,50 +101,9 @@ impl Mergeable for StartPhase {
     }
 }
 
-/// Fills in the `"..."`'s or `"@auto"`'s in `replacer` with the values from the `original`
-///
-/// ```
-/// let arr = fill_auto_in_vec(
-///   Some(vec!["a", "b", "c"]),
-///   Some(vec!["x", "...", "z"])
-/// );
-/// assert_eq!(Some(vec!["x", "...", "a", "b", "c", "z"]), arr);
-/// ```
-fn fill_auto_in_vec(
-    original: Option<Vec<String>>,
-    replacer: Option<Vec<String>>,
-) -> Option<Vec<String>> {
-    if let Some(replacer) = replacer {
-        let original = original.unwrap_or_default();
-        let modified = replacer
-            .into_iter()
-            .flat_map(|x| {
-                let v = x.clone();
-                if v == *"@auto" || v == *"..." {
-                    let mut fill = vec![v];
-                    fill.append(&mut original.clone());
-                    fill
-                } else {
-                    vec![x]
-                }
-            })
-            .collect::<Vec<_>>();
-
-        Some(modified)
-    } else {
-        original
-    }
-}
-
 #[cfg(test)]
 mod test {
     use super::*;
-
-    fn vs(v: Vec<&str>) -> Vec<String> {
-        v.into_iter()
-            .map(std::string::ToString::to_string)
-            .collect()
-    }
 
     #[test]
     fn test_merge_plan() {
@@ -242,26 +202,6 @@ mod test {
             )
             .unwrap(),
             merged
-        );
-    }
-
-    #[test]
-    fn test_fill_auto_in_vec() {
-        assert_eq!(
-            vec!["x", "...", "z"],
-            fill_auto_in_vec(None, Some(vs(vec!["x", "...", "z"]))).unwrap()
-        );
-        assert_eq!(
-            vec!["a", "b", "c"],
-            fill_auto_in_vec(Some(vs(vec!["a", "b", "c"])), None).unwrap()
-        );
-        assert_eq!(
-            vec!["x", "...", "a", "b", "c", "z"],
-            fill_auto_in_vec(
-                Some(vs(vec!["a", "b", "c"])),
-                Some(vs(vec!["x", "...", "z"]))
-            )
-            .unwrap()
         );
     }
 }
